Tweak spoiler command

+++ src/cmds/spoiler.py
@@ -1,33 +1,38 @@
-from typing import Union
 import discord
-from discord.ext import commands
-from discord.commands import Option
 from discord.commands.context import ApplicationContext
-from src.noahbot import bot
-from src.conf import SlashPerms, PrefixPerms, ChannelIDs,GUILD_ID
+from discord.ext import commands
+
 from src.cmds._proxy_helpers import Reply
+from src.conf import ChannelIDs, GUILD_ID
+from src.noahbot import bot
 
 
 def name():
     return 'spoiler'
 
+
 def description():
     return 'Add the URL which has spoiler link'
+
 
 async def perform_action(ctx: ApplicationContext, reply, url: str):
     if len(url) == 0:
         await ctx.send('Please provide the spoiler URL.')
         return
-    
-    ctx.message.delete()
-    embed = discord.Embed(title="Error: Spoiler Report.", color=0xFF0000)
-    await bot.get_channel(ChannelIDs.BOT_LOGS).send(embed=embed(ctx, f"{ctx.author.mention} has submitted a spoiler report"+ url))
-    await reply(ctx, "Thanks for the reporting the spoiler")
+
+    if ctx.guild and reply == Reply.prefix:
+        await ctx.message.delete()
+
+    embed = discord.Embed(title="Spoiler Report", color=0xb98700)
+    embed.add_field(name=f"{ctx.author} has submitted a spoiler.", value=f"URL: <{url}>", inline=False)
+
+    await bot.get_channel(ChannelIDs.SPOILER_CHAN).send(embed=embed)
+    await reply(ctx, "Thanks for the reporting the spoiler", ephemeral=True, delete_after=15)
 
 
 @bot.slash_command(guild_ids=[GUILD_ID], name=name(), description=description())
-async def action_slash(ctx: ApplicationContext, reply, url: str):
-    await perform_action(ctx, reply, url)
+async def action_slash(ctx: ApplicationContext, url: str):
+    await perform_action(ctx, Reply.slash, url)
 
 
 @commands.command(name=name(), help=description())
@@ -35,6 +40,5 @@ async def action_prefix(ctx: ApplicationContext, url: str):
     await perform_action(ctx, Reply.prefix, url)
 
 
-
-def setup():
-    le_bot = bot.add_command(perform_action)
+def setup(le_bot):
+    le_bot.add_command(action_prefix)
